Added a config file to visualStimToolbox
The default values of the screens are Pc specific, but the VSconfig file relates to VStim.m only. So I copied the section refering to the config file and modified it to work in visualStimToolbox.m
Works for setting screens, but the location and MO might not work with other properties.