fix SyncSquare luminosity
previously, changing the value of SyncSquareLuminosity in VStim from the defualt 255 resulted in a change of both luminosity and transparency, due to how the ON condition was constructed.
It is now changed and any value of 0-255 can be displayed, while the transparency remains the required 255.
This is a patch up solution, though. Probably should redefine the process of constructing the SyncSquare in the future, to allow for transparency modification, shoudl one need to.

GuyH

Co-Authored-By: guyhyman <guyhyman@users.noreply.github.com>
Co-Authored-By: Mark Josef <mcphage@gmail.com>