Create VS_fullFieldFlash_retarder.m
Created an initial stimulation for applying voltages and functions in the new RIGOL AWG. Currently, it is a standard fullFieldFlash, with added commands at the initiation and following each instance of the "sendTTL" function.
Before the main loop, it opens the visa object used to control the AWG. It then set the carrier waveform parameters.
Next, after every use of sendTTL for bit 2 (trial gating), the amplitude of the waveform is changed accordingly (TTL ON = voltage for 90deg rotation and the opposite).
When sendTTL is used for bit 1 (start/end of stimualtion), the Output of the AWG is turned on and off, accordingly.

Lastly, this is very temporary.
Future additions:
- remove screen sections
- add variable angles
- change GUI for inputing the parameter there
- wavelength needs to be accounted for somehow

Co-Authored-By: guyhyman <60943471+guyhyman@users.noreply.github.com>