workaround 2020 change
As of 2020a, readtable has some behavioral changes.
In my case, it started reading the xlsx file 40 rows after the header. Don't know why exactly why.
According to the docs, setting the 'format' property to auto reverts the behavior of readtable back. This workaround wroked for me. Since I can't be sure this works for earlier versions, I conditioned the workaround to 2020 and after.