Changed condition
I found a built-in function that compares the version number to a given version. Now the original way readtable was called is used when the condition applies (i.e. current version is less than 9.8) and the workaround is used when not.
The change happened in version 2020a, which is version number 9.8. Since this is a built-in function of matlab and we're dealing with version number rather release name, this should be a robust solution.