Analog data in timeSeriesViewer + modify plots in sleepAnalysis
add the possibility to look at analog data. Feature added by modifying the activityTrace class. But will create problems when averaging trials. Another option is to integrate with electrode data already in time series viewer and make the separation in AVPlotData based on channel name.
In addition, minor changes in a