new way to find VSFile
Added a new way for MEAAnalysis.getVSParam to find the VSFile in the main recording directory. Current method requires a specific special folder, which is easily missed. This method relies on the VSMetaData variable, which the VSFile always contains and is constantly named as VSMetaData.
Future editions will include the ability to search all subfolders as well and, as MArk suggested for his methods, use the date and time as an additional or replacement reference.